Refactor for better separation of concerns - move toggle_status to Todo context
