+++ todos/test/todos_web/controllers/item_controller_test.exs
@@ -128,19 +128,8 @@ defmodule TodosWeb.ItemControllerTest do
     end
   end
 
-  describe "toggle_status updates the status of an item" do
+  describe "toggle updates the status of an item" do
     setup [:create_item]
-
-    test "toggle_status/1 item.status 1 > 0", %{item: item} do
-      assert item.status == 0
-
-      # first toggle
-      toggled_item = %{item | status: TodosWeb.ItemController.toggle_status(item)}
-      assert toggled_item.status == 1
-
-      # second toggle
-      assert TodosWeb.ItemController.toggle_status(toggled_item) == 0
-    end
 
     test "toggle/2 (toggle endpoint) updates an item.status 0 > 1", %{conn: conn, item: item} do
       assert item.status == 0
